# webapi.event+/Controllers/ComentariosEventoController.cs
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Azure.CognitiveServices.ContentModerator;
using Microsoft.IdentityModel.Tokens;
using System.Text;
using webapi.event_.Domains;
using webapi.event_.Repositories;

namespace webapi.event_.Controllers
{
	[Route("api/[controller]")]
	[ApiController]
	[Produces("application/json")]
	public class ComentariosEventoController : ControllerBase
	{
		ComentariosEventoRepository comentario = new ComentariosEventoRepository();

		//armazena dadis di servico da Api externa (IA -Azure)
		private readonly ContentModeratorClient _contentModeratorClient;

		/// <summary>
		/// Construtor que recebe os dados necessarios para acesso ao servico externo
		/// </summary>
		/// <param name="contentModeratorClient">objeto do tipo ContentModeratorClinet</param>
		public ComentariosEventoController(ContentModeratorClient contentModeratorClient)
		{
			_contentModeratorClient = contentModeratorClient;
		}

		[HttpPost("ComentarioIA")]
		public async Task<IActionResult> PostIa(ComentariosEvento novoComentario)
		{
			try
			{
				//if(comentario.Descricao !=  null || comentario.Descricao == "")
				//if (string.IsNullOrEmpty(comentario.Descricao))
				//if ( (comentario.Descricao).IsNullOrEmpty()) 

				if (string.IsNullOrEmpty(novoComentario.Descricao))
				{
					return BadRequest("A descrição do comentario não pode estar vazio ou nulo.");
				}

				using var stream = new MemoryStream(Encoding.UTF8.GetBytes(novoComentario.Descricao));

				var moderationResult = await _contentModeratorClient.TextModeration
					.ScreenTextAsync("text/plain", stream, "por", false, false, null, true);

				if(moderationResult.Terms != null)
				{
					novoComentario.Exibe = false;

					comentario.Cadastrar(novoComentario);

				}
				else
				{
					novoComentario.Exibe = true;

					comentario.Cadastrar(novoComentario);
				}
				return StatusCode(201, novoComentario);
			}
			catch (Exception e)
			{
				return BadRequest(e.Message);
			}
		}



		[HttpGet]
		public IActionResult Get()
		{
			try
			{
				return Ok(comentario.Listar());
			}
			catch (Exception e)
			{

				return BadRequest(e.Message);
			}
		}
		
		
		[HttpGet("ListarSomenteExibe")]
		public IActionResult GetShow(Guid id)
		{
			try
			{
				return Ok(comentario.ListarSomenteExibe(id));
			}
			catch (Exception e)
			{

				return BadRequest(e.Message);
			}
		}

		[HttpGet("BuscarPorIdUsuario")]
		public IActionResult GetByIdUser(Guid idUsuario, Guid idEvento)
		{
			try
			{
				return Ok(comentario.BuscarPorIdUsuario(idUsuario, idEvento));
			}
			catch (Exception)
			{

				throw;
			}
		}

		[HttpPost]
		public IActionResult Post(ComentariosEvento novoComentario)
		{
			try
			{
				comentario.Cadastrar(novoComentario);

				return StatusCode(201, novoComentario);
			}
			catch (Exception e)
			{

				return BadRequest(e.Message);
			}
		}

		[HttpDelete("{id}")]
		public IActionResult Delete(Guid id)
		{
			try
			{
				comentario.Deletar(id);

				return NoContent();
			}
			catch (Exception e)
			{

				return BadRequest(e.Message);
			}
		}

	}
}